init wayland support.
The display server based on X or Wayland, is now a choice of different profile in Tizen 3.0.
Basically and consistently, two macros were used "with wayland" and "with x".
Below summarize the combination of the macros:

| wayland |  x  | meaning
|---------------------------
|   0     |  1  | pure X11 platform(no wayland)
|   1     |  0  | pure wayland platform (no X11)
|   1     |  1  | wayland but X compatibility
|   0     |  0  | no X and no wayland

This method unifies the meaning and usage. Deploy this method to capi-system-info package.
Currently we need "pure wayland platform (no X11)" status,so "with wayland" and "with x" are
defined into "1" and "0" in build.conf file.If you need other status of above illustration,
please change values of "with x" and "with wayland" in your build.conf file.

Change-Id: If8c3b9e099d60fc892140fe480fe6b52e8a66d6a
Signed-off-by: wangfei <feix.w.wang@intel.com>